+++ pkg/alertmanager/statefulset_test.go
@@ -28,7 +28,8 @@ import (
 
 var (
 	defaultTestConfig = Config{
-		ConfigReloaderImage: "quay.io/coreos/configmap-reload:latest",
+		ConfigReloaderImage:          "quay.io/coreos/configmap-reload:latest",
+		AlertmanagerDefaultBaseImage: "quay.io/prometheus/alertmanager",
 	}
 )
 
@@ -391,6 +392,24 @@ func TestAdditionalSecretsMounted(t *testing.T) {
 
 	if !(secret1Found && secret2Found) {
 		t.Fatal("Additional secrets were not found.")
+	}
+}
+
+func TestTagAndVersion(t *testing.T) {
+	sset, err := makeStatefulSet(&monitoringv1.Alertmanager{
+		Spec: monitoringv1.AlertmanagerSpec{
+			Tag:     "my-unrelated-tag",
+			Version: "v0.15.0",
+		},
+	}, nil, defaultTestConfig)
+	if err != nil {
+		t.Fatalf("Unexpected error while making StatefulSet: %v", err)
+	}
+
+	image := sset.Spec.Template.Spec.Containers[0].Image
+	expected := "quay.io/prometheus/alertmanager:my-unrelated-tag"
+	if image != expected {
+		t.Fatalf("Unexpected container image.\n\nExpected: %s\n\nGot: %s", expected, image)
 	}
 }