+++ test/e2e/prometheus_test.go
@@ -22,6 +22,7 @@ import (
 	"log"
 	"reflect"
 	"sort"
+	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -333,6 +334,8 @@ func TestPrometheusAdditionalScrapeConfig(t *testing.T) {
 }
 
 func TestPrometheusAdditionalAlertManagerConfig(t *testing.T) {
+	t.Parallel()
+
 	ctx := framework.NewTestCtx(t)
 	defer ctx.Cleanup(t)
 	ns := ctx.CreateNamespace(t, framework.KubeClient)
@@ -468,50 +471,6 @@ func TestPrometheusReloadRules(t *testing.T) {
 	}
 }
 
-// With Prometheus Operator v0.20.0 the 'RuleSelector' field in the Prometheus
-// CRD Spec is deprecated. We need to ensure to still support it until the field
-// is removed. Any value in 'RuleSelector' should just be copied to the new
-// field 'RuleFileSelector'.
-func TestPrometheusDeprecatedRuleSelectorField(t *testing.T) {
-	t.Parallel()
-
-	ctx := framework.NewTestCtx(t)
-	defer ctx.Cleanup(t)
-	ns := ctx.CreateNamespace(t, framework.KubeClient)
-	ctx.SetupPrometheusRBAC(t, ns, framework.KubeClient)
-
-	name := "test"
-	firtAlertName := "firstAlert"
-
-	_, err := framework.MakeAndCreateFiringRule(ns, name, firtAlertName)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	p := framework.MakeBasicPrometheus(ns, name, name, 1)
-	p.Spec.EvaluationInterval = "1s"
-	p.Spec.RuleSelector = &metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"role": "rulefile",
-		},
-	}
-	if err := framework.CreatePrometheusAndWaitUntilReady(ns, p); err != nil {
-		t.Fatal(err)
-	}
-
-	pSVC := framework.MakePrometheusService(p.Name, "not-relevant", v1.ServiceTypeClusterIP)
-	if finalizerFn, err := testFramework.CreateServiceAndWaitUntilReady(framework.KubeClient, ns, pSVC); err != nil {
-		t.Fatal(errors.Wrap(err, "creating Prometheus service failed"))
-	} else {
-		ctx.AddFinalizerFn(finalizerFn)
-	}
-
-	err = framework.WaitForPrometheusFiringAlert(p.Namespace, pSVC.Name, firtAlertName)
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
 func TestPrometheusRuleConfigMapMigration(t *testing.T) {
 	t.Parallel()
 
@@ -541,7 +500,9 @@ groups:
 `, alertName),
 		},
 	}
-	framework.KubeClient.CoreV1().ConfigMaps(ns).Create(&cm)
+	if _, err := framework.KubeClient.CoreV1().ConfigMaps(ns).Create(&cm); err != nil {
+		t.Fatalf("failed to create legacy rule config map: %v", err.Error())
+	}
 
 	p := framework.MakeBasicPrometheus(ns, name, name, 1)
 	p.Spec.RuleSelector = &metav1.LabelSelector{
@@ -569,7 +530,7 @@ groups:
 	}
 }
 
-func TestPrometheusMultipleRuleFilesSameNS(t *testing.T) {
+func TestPrometheusMultiplePrometheusRulesSameNS(t *testing.T) {
 	t.Parallel()
 
 	ctx := framework.NewTestCtx(t)
@@ -608,7 +569,7 @@ func TestPrometheusMultipleRuleFilesSameNS(t *testing.T) {
 	}
 }
 
-func TestPrometheusMultipleRuleFilesDifferentNS(t *testing.T) {
+func TestPrometheusMultiplePrometheusRulesDifferentNS(t *testing.T) {
 	t.Parallel()
 
 	ctx := framework.NewTestCtx(t)
@@ -627,7 +588,10 @@ func TestPrometheusMultipleRuleFilesDifferentNS(t *testing.T) {
 	ruleFilesNamespaceSelector := map[string]string{"prometheus": rootNS}
 
 	for _, file := range ruleFiles {
-		testFramework.AddLabelsToNamespace(framework.KubeClient, file.ns, ruleFilesNamespaceSelector)
+		err := testFramework.AddLabelsToNamespace(framework.KubeClient, file.ns, ruleFilesNamespaceSelector)
+		if err != nil {
+			t.Fatal(err)
+		}
 	}
 
 	for _, file := range ruleFiles {
@@ -659,6 +623,104 @@ func TestPrometheusMultipleRuleFilesDifferentNS(t *testing.T) {
 			t.Fatal(err)
 		}
 	}
+}
+
+func TestPrometheusRulesExceedingConfigMapLimit(t *testing.T) {
+	t.Parallel()
+
+	ctx := framework.NewTestCtx(t)
+	defer ctx.Cleanup(t)
+	ns := ctx.CreateNamespace(t, framework.KubeClient)
+	ctx.SetupPrometheusRBAC(t, ns, framework.KubeClient)
+
+	prometheusRules := []monitoringv1.PrometheusRule{}
+	for i := 0; i < 2; i++ {
+		rule := generateHugePrometheusRule(ns, strconv.Itoa(i))
+		err := framework.CreateRule(ns, rule)
+		if err != nil {
+			t.Fatal(err)
+		}
+		prometheusRules = append(prometheusRules, rule)
+	}
+
+	name := "test"
+
+	p := framework.MakeBasicPrometheus(ns, name, name, 1)
+	p.Spec.EvaluationInterval = "1s"
+	if err := framework.CreatePrometheusAndWaitUntilReady(ns, p); err != nil {
+		t.Fatal(err)
+	}
+
+	defer func() {
+		if t.Failed() {
+			if err := framework.PrintPodLogs(ns, "prometheus-"+p.Name+"-0"); err != nil {
+				t.Fatal(err)
+			}
+		}
+	}()
+
+	pSVC := framework.MakePrometheusService(p.Name, "not-relevant", v1.ServiceTypeClusterIP)
+	if finalizerFn, err := testFramework.CreateServiceAndWaitUntilReady(framework.KubeClient, ns, pSVC); err != nil {
+		t.Fatal(errors.Wrap(err, "creating Prometheus service failed"))
+	} else {
+		ctx.AddFinalizerFn(finalizerFn)
+	}
+
+	for i := range prometheusRules {
+		_, err := framework.WaitForConfigMapExist(ns, "prometheus-"+p.Name+"-rulefiles-"+strconv.Itoa(i))
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
+
+	// Make sure both rule files ended up in the Prometheus Pod
+	for i := range prometheusRules {
+		err := framework.WaitForPrometheusFiringAlert(ns, pSVC.Name, "my-alert-"+strconv.Itoa(i))
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
+
+	err := framework.DeleteRule(ns, prometheusRules[1].Name)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	_, err = framework.WaitForConfigMapExist(ns, "prometheus-"+p.Name+"-rulefiles-0")
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = framework.WaitForConfigMapNotExist(ns, "prometheus-"+p.Name+"-rulefiles-1")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = framework.WaitForPrometheusFiringAlert(ns, pSVC.Name, "my-alert-0")
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+// generateHugePrometheusRule returns a Prometheus rule instance that would fill
+// more than half of the space of a Kubernetes config map.
+func generateHugePrometheusRule(ns, identifier string) monitoringv1.PrometheusRule {
+	alertName := "my-alert"
+	groups := []monitoringv1.RuleGroup{
+		monitoringv1.RuleGroup{
+			Name:  alertName,
+			Rules: []monitoringv1.Rule{},
+		},
+	}
+	// One rule marshaled as yaml is ~34 bytes long, the max is ~524288 bytes.
+	for i := 0; i < 12000; i++ {
+		groups[0].Rules = append(groups[0].Rules, monitoringv1.Rule{
+			Alert: alertName + "-" + identifier,
+			Expr:  "vector(1)",
+		})
+	}
+	rule := framework.MakeBasicRule(ns, "prometheus-rule-"+identifier, groups)
+
+	return rule
 }
 
 // Make sure the Prometheus operator only updates the Prometheus config secret
@@ -703,9 +765,12 @@ func TestPrometheusOnlyUpdatedOnRelevantChanges(t *testing.T) {
 					KubeClient.
 					CoreV1().
 					ConfigMaps(ns).
-					Get("prometheus-"+prometheusName+"-rulefiles", metav1.GetOptions{})
+					Get("prometheus-"+prometheusName+"-rulefiles-0", metav1.GetOptions{})
 			},
-			MaxExpectedChanges: 1,
+			// The Prometheus Operator first creates the config map for the
+			// given Prometheus stateful set and then updates it with the matching
+			// Prometheus rules.
+			MaxExpectedChanges: 2,
 		},
 		{
 			Name: "configurationSecret",
@@ -773,7 +838,24 @@ func TestPrometheusOnlyUpdatedOnRelevantChanges(t *testing.T) {
 		}
 	}()
 
+	alertName := "my-alert"
+	if _, err := framework.MakeAndCreateFiringRule(ns, "my-prometheus-rule", alertName); err != nil {
+		t.Fatal(err)
+	}
+
 	if err := framework.CreatePrometheusAndWaitUntilReady(ns, prometheus); err != nil {
+		t.Fatal(err)
+	}
+
+	pSVC := framework.MakePrometheusService(prometheus.Name, "not-relevant", v1.ServiceTypeClusterIP)
+	if finalizerFn, err := testFramework.CreateServiceAndWaitUntilReady(framework.KubeClient, ns, pSVC); err != nil {
+		t.Fatal(errors.Wrap(err, "creating Prometheus service failed"))
+	} else {
+		testCTX.AddFinalizerFn(finalizerFn)
+	}
+
+	err := framework.WaitForPrometheusFiringAlert(prometheus.Namespace, pSVC.Name, alertName)
+	if err != nil {
 		t.Fatal(err)
 	}
 
@@ -820,7 +902,7 @@ func TestPrometheusWhenDeleteCRDCleanUpViaOwnerReference(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	configMapName := fmt.Sprintf("prometheus-%v-rulefiles", p.Name)
+	configMapName := fmt.Sprintf("prometheus-%v-rulefiles-0", p.Name)
 
 	_, err := framework.WaitForConfigMapExist(ns, configMapName)
 	if err != nil {