+++ test/framework/prometheus_rule.go
@@ -92,3 +92,12 @@ func (f *Framework) UpdateRule(ns string, ar monitoringv1.PrometheusRule) error 
 
 	return nil
 }
+
+func (f *Framework) DeleteRule(ns string, r string) error {
+	err := f.MonClientV1.PrometheusRules(ns).Delete(r, &metav1.DeleteOptions{})
+	if err != nil {
+		return fmt.Errorf("deleteing %v Prometheus rule in namespace %v failed: %v", r, ns, err.Error())
+	}
+
+	return nil
+}